Refactor zenoh diagram operation registration
